==> development/01 main.ipynb <==
import pandas as pd
from sklearn.model_selection import train_test_split

df = pd.read_csv("train_data.csv", index_col=0).drop_duplicates()

print(len(df))

df.head()

df["Тип оборудования"].value_counts()

df["Точка отказа"].value_counts()

train, test = train_test_split(df, stratify=df["Тип оборудования"], train_size=0.7)

train["Тип оборудования"].value_counts()

train.to_csv("train.csv")

test["Тип оборудования"].value_counts()

test.to_csv("test.csv")

train.head()

train["Тема"].to_list()
