==> development/02 datav2.ipynb <==
import pandas as pd
import csv

def get_text_data(dataframe, file_name="train.csv"):
    texts, labels = [], []
    for i, row in dataframe.iterrows():
        texts.append(row["Тема"] + " [SEP] " + row["Описание"])
        labels.append(row["Тип оборудования"])
    # print(texts)
    # Объединяем списки в записи
    data = zip(texts, labels)
    # Сохраняем в CSV файл с разделителем \t
    with open(f'datav2/{file_name}', 'w', newline='') as file:
        writer = csv.writer(file, delimiter='\t')
        writer.writerows(data)
        

test = pd.concat([pd.read_csv("test.csv", index_col=0), pd.read_csv("train.csv", index_col=0)]).reset_index(drop=True)
train = pd.read_csv("new_data_v2_increased.csv")

test["Тип оборудования"].value_counts()

train["Тип оборудования"].value_counts()

get_text_data(train, "train.csv")
get_text_data(test, "dev.csv")
get_text_data(test, "test.csv")

train["Тема"].to_list()

type(train["Тема"])
