==> development/03 run_expv2.ipynb <==
model_names = [ 'deepvk/USER-bge-m3', 'deepvk/USER-base',
               'intfloat/multilingual-e5-large-instruct', 'intfloat/multilingual-e5-large',
               'sentence-transformers/paraphrase-multilingual-mpnet-base-v2', 'intfloat/multilingual-e5-base', 'intfloat/multilingual-e5-small', 
               ]
script_name = "main.py"
for model in model_names:
    data_folder='./datav2/'
    experiment = 'Тип_оборудования_data_v2.0_expanded_data'
    model_name = model
    learning_rate = 5.0e-5
    mini_batch_size = 8
    max_epochs = 5
    
    command = f"python {script_name} --experiment {experiment} --data_folder {data_folder} --model_name {model_name} --learning_rate {learning_rate} --mini_batch_size {mini_batch_size} --max_epochs {max_epochs}"
    !{command}

model_names = [ 'deepvk/USER-bge-m3', 'deepvk/USER-base',
               'intfloat/multilingual-e5-large-instruct', 'intfloat/multilingual-e5-large',
               'sentence-transformers/paraphrase-multilingual-mpnet-base-v2', 'intfloat/multilingual-e5-base',
               'sergeyzh/rubert-tiny-turbo', 'intfloat/multilingual-e5-small', 
               "FacebookAI/xlm-roberta-large"
               ]
script_name = "main.py"
for model in model_names:
    data_folder='./datav2/'
    experiment = 'Тип_оборудования_data_v2.0_expanded_data'
    model_name = model
    learning_rate = 5.0e-5
    mini_batch_size = 8
    max_epochs = 5
    
    command = f"python {script_name} --experiment {experiment} --data_folder {data_folder} --model_name {model_name} --learning_rate {learning_rate} --mini_batch_size {mini_batch_size} --max_epochs {max_epochs}"
    !{command}

#'deepvk/USER-bge-m3', 'deepvk/USER-base',
               'intfloat/multilingual-e5-large-instruct', 'intfloat/multilingual-e5-large',
               'sentence-transformers/paraphrase-multilingual-mpnet-base-v2', 'intfloat/multilingual-e5-base',
               'sergeyzh/rubert-tiny-turbo', 'intfloat/multilingual-e5-small', 
               "FacebookAI/xlm-roberta-large", "yandex/RuLeanALBERT",

import pandas as pd

pd.read_csv("data/train.csv", sep='\t')
