==> development/05 model copy.ipynb <==
from transformers import AutoTokenizer, AutoModel, BertForSequenceClassification
import torch
from torch.utils.data import Dataset, DataLoader
import pandas as pd
from transformers import get_scheduler, AdamW

from typing import Dict

device = torch.device("cuda" if torch.cuda.is_available() else "cpu")

train = pd.read_csv("train.csv", index_col=0)
test = pd.read_csv("test.csv", index_col=0)
new_data = pd.read_csv("new_data.csv")

train["Тип оборудования"].unique()

label2id, id2label = dict(), dict()
for id_, class_ in enumerate(train["Тип оборудования"].unique()):
    label2id[class_] = id_
    id2label[id_] = class_    

id2label

model = BertForSequenceClassification.from_pretrained("blinoff/roberta-base-russian-v0", num_labels=len(label2id), id2label=id2label, label2id=label2id).to(device)
tokenizer = AutoTokenizer.from_pretrained("blinoff/roberta-base-russian-v0")

class CustomDataset(Dataset):
    def __init__(self, dataframe: pd.DataFrame,
                tokenizer: AutoTokenizer,
                label2id: Dict[str, int]):
        self.tokenizer = tokenizer
        self.max_length: int = tokenizer.model_max_length
        self.sep_token: str = tokenizer.sep_token

        dataframe["text"] = dataframe["Тема"] + f" {self.sep_token} " + dataframe["Описание"]
        self.texts = dataframe["text"].to_list()
        self.labels = [label2id[label] for label in dataframe["Тип оборудования"]]
    
    def __len__(self):
        return len(self.texts)
    
    def __getitem__(self, index: int):
        text :str = self.texts[index]
        label: int = self.labels[index]

        encoding = self.tokenizer.encode_plus(
            text,
            add_special_tokens=True,
            max_length=self.max_length,
            return_token_type_ids=False,
            padding='max_length',
            truncation=True,
            return_attention_mask=True,
            # Return PyTorch tensors
            return_tensors='pt',
        )

        return {
            'input_ids': encoding['input_ids'].flatten(),
            'attention_mask': encoding['attention_mask'].flatten(),
            'labels': torch.tensor(label, dtype=torch.long)
            }

from sklearn.metrics import f1_score, classification_report


def train_model(model, data_loader, optimizer, lr_scheduler, device):
    model = model.train()
    total_loss = 0
    for batch in data_loader:
        input_ids = batch['input_ids'].to(device)
        attention_mask = batch['attention_mask'].to(device)
        labels = batch['labels'].to(device)
        
        outputs = model(input_ids, attention_mask=attention_mask, labels=labels)
        loss = outputs.loss
        
        optimizer.zero_grad()
        loss.backward()
        optimizer.step()
        lr_scheduler.step()
        
        total_loss += loss.item()
    return total_loss / len(data_loader)

def evaluate_model(model, data_loader, device, last_epoch=False):
    model = model.eval()
    total_loss = 0
    all_predictions = []
    all_true_labels = []

    with torch.no_grad():
        for batch in data_loader:
            input_ids = batch['input_ids'].to(device)
            attention_mask = batch['attention_mask'].to(device)
            labels = batch['labels'].to(device)

            outputs = model(input_ids, attention_mask=attention_mask, labels=labels)
            loss = outputs.loss
            total_loss += loss.item()

            logits = outputs.logits
            predictions = torch.argmax(logits, dim=1)
            all_predictions.extend(predictions.cpu().numpy())
            all_true_labels.extend(labels.cpu().numpy())

    # Calculate F1 Score
    f1 = f1_score(all_true_labels, all_predictions, average='weighted')  # choose average type as 'macro' or 'micro' if needed
    if last_epoch:
        print(classification_report(all_true_labels, all_predictions, target_names=label2id.keys()))
    return total_loss / len(data_loader), f1


def calculate_model(model, data_loader, device, last_epoch=False):
    model = model.eval()
    total_loss = 0
    all_predictions = []
    all_true_labels = []

    with torch.no_grad():
        for batch in data_loader:
            input_ids = batch['input_ids'].to(device)
            attention_mask = batch['attention_mask'].to(device)
            labels = batch['labels'].to(device)

            outputs = model(input_ids, attention_mask=attention_mask, labels=labels)
            loss = outputs.loss
            total_loss += loss.item()

            logits = outputs.logits
            predictions = torch.argmax(logits, dim=1)
            all_predictions.extend(predictions.cpu().numpy())
            all_true_labels.extend(labels.cpu().numpy())

    # Calculate F1 Score

    cl = classification_report(all_true_labels, all_predictions, target_names=label2id.keys())
    return total_loss / len(data_loader), cl

train_dataset = CustomDataset(train, tokenizer, label2id)
test_dataset = CustomDataset(test, tokenizer, label2id) 
new_dataset = CustomDataset(new_data, tokenizer, label2id)

train_loader = DataLoader(train_dataset, batch_size=4, shuffle=True)
test_loader = DataLoader(test_dataset, batch_size=4, shuffle=True)
new_loader = DataLoader(new_dataset, batch_size=4, shuffle=True)

optimizer = AdamW(model.parameters(), lr=1e-5, eps = 1e-8)
num_epochs = 3
num_training_steps = num_epochs * len(train_loader)
lr_scheduler = get_scheduler(
    name="linear",
    optimizer=optimizer,
    num_warmup_steps=0,
    num_training_steps=num_training_steps
)

model = model.train()
for epoch in range(num_epochs):
    train_loss = train_model(model, train_loader, optimizer, lr_scheduler, device)
    val_loss, f1 = evaluate_model(model, test_loader, device, epoch==num_epochs-1)
    print(f"Epoch: {epoch}, Train Loss: {train_loss:.4f}, Validation Loss: {val_loss:.4f}, Validation F1: {f1:.4f}")

val_loss, f1 = evaluate_model(model, test_loader, device, True)

def batch_predict(texts, model, tokenizer, device):
    # Ensure the model is in evaluation mode
    model.eval()
    
    # Tokenize the batch of texts
    encodings = tokenizer(texts, max_length=tokenizer.model_max_length, truncation=True, padding=True, return_tensors="pt")
    
    # Move the tensors to the appropriate device
    input_ids = encodings['input_ids'].to(device)
    attention_mask = encodings['attention_mask'].to(device)
    
    # Perform inference
    with torch.no_grad():
        outputs = model(input_ids, attention_mask=attention_mask)
        logits = outputs.logits
    
    # Convert logits to probabilities and then to actual predictions
    probabilities = torch.nn.functional.softmax(logits, dim=1)
    predictions = torch.argmax(probabilities, dim=1).cpu().tolist()
    
    return [id2label[prediction] for prediction in predictions]

batch_predict(["Проблемы с компом, сервер SERVER", "открыть кейс по неисправному диску на dd:  dd203..ru  Enclosure=3:Disk=49  открыть кейс по неисправному диску на dd:  dd203..ru  Enclosure=3:Disk=49"], model, tokenizer, device)

_, cl = calculate_model(model, new_loader, device)

print(cl)

train_loader.dataset

print(cl)

new_loader.dataset.iloc[924]
