==> development/06 new_data_v2.ipynb <==
import json
import pandas as pd

with open('new_data_gpt4o_v2.jsonl') as f:
    data = [json.loads(line) for line in f]

result = []
for row in data:
    for val in row["Обращения"].values():
        result.append({
            'Тема': val['Тема'],
            'Описание': val['Описание'],
            'Тип оборудования': row['Тип оборудования'],
            'Точка отказа': row['Точка отказа']})

df = pd.DataFrame(result)

df["Точка отказа"].value_counts()

df_prev = pd.read_csv("new_data_v2.csv", index_col=0).drop(columns=["Серийный номер"])

# df_prev = df_prev[df_prev["Точка отказа"] != "Консультация"]
# df = df[df["Точка отказа"] != "Консультация"]

df_prev["Точка отказа"] = df_prev["Точка отказа"].apply(lambda x: x.strip())
df["Точка отказа"] = df["Точка отказа"].apply(lambda x: x.strip())

len(df), len(df_prev)

df.head()

df_prev.head()

df_final = pd.concat([df, df_prev])

assert len(df_final) == len(df) + len(df_prev)

df_final = df_final.reset_index(drop=True)

df_final.head()

df_final.to_csv("new_data_v2_increased.csv", index=False)

df_ = pd.read_csv("new_data_v2.csv", index_col=0)

df_.sample(frac=1)[:100].to_csv('submit_sample.csv', sep=",", index=False)

data = pd.read_csv('submit_sample.csv', sep=",")

data["Тема"]

df.loc[1522]

df[df["Тема"]=="FW: Не работает экран ноутбука"]["Описание"].tolist()

sub = pd.read_csv("submit.csv")

!nvidia-smi
